requirements.txt
fastapi>=0.121.0
uvicorn[standard]>=0.38.0
pydantic>=2.12.4
pydantic-settings>=2.11.0
sqlalchemy>=2.0.44
sqlmodel>=0.0.27
aiosqlite>=0.21.0
httpx>=0.28.1
pytest>=8.4.2
pytest-asyncio>=1.2.0
mypy>=1.18.2
ruff>=0.14.4
bandit>=1.8.6
opentelemetry-sdk>=1.38.0
opentelemetry-instrumentation-fastapi>=0.59b0
python-jose>=3.5.0
python-multipart>=0.0.20
boto3>=1.40.68
types-requests>=2.32.4.20250913
